pyproject.toml
[project]
name = "manhwa-russifier"
version = "1.0.0"
description = "A complete pipeline for automatic russification of manhwa: text detection, OCR, translation, replacement, and rendering."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ultralytics>=8.0.0,<9.0.0",
    "easyocr>=1.7.0,<2.0.0",
	"googletrans>=4.0.0,<5.0.0",
    "torch>=2.0.0,<3.0.0",
    "torchvision>=0.15.0,<1.0.0",
    "opencv-python>=4.8.0,<5.0.0",
    "numpy>=1.24.0,<3.0.0",
    "pillow>=10.0.0,<11.0.0",
    "py7zr>=0.20.0,<1.0.0",
    "git+https://github.com/KvaytG/caption-forge.git@d35e91c2e59cbea40322469e032c02703df869fc",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<8.0.0",
    "black>=24.0.0,<25.0.0",
    "mypy>=1.8.0,<2.0.0",
    "flake8>=7.0.0,<8.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}